Remove condition from recipe
It seems there's some confusion in terms of overriding steampowered's recipes. So the recipes will still be here, but the eventual goal is to relocate them to their own datapack which can then be uploaded to curseforge alongside the main mod, Users can be instructed to use that datapack to override the recipes.